Reorganize into src
